Partial load feature implemented in CSLogger module:
There is a checkbox to allow or not the partial loading of files, it means if there are missing files for an id, the partial feature allows to submit the remaining. If partial is false the complete task is discard.

The other types of responses different to media files are not considered for missing_media array in CSLogger.vue